% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_memory_span.R
\name{measure_memory_span}
\alias{measure_memory_span}
\title{Measure Memory Span}
\usage{
measure_memory_span(
  span_length = 1,
  type = c("digit", "letter"),
  order = c("forward", "reverse", "random")
)
}
\arguments{
\item{span_length}{the length of the first sequence of digits or letters.
Increments by 1 after each correct response.}

\item{type}{measure "digit" (default) or "letter" span?}

\item{order}{should the response be returned in "forward" order (default),
"reverse" order, or "random" (i.e., switch randomly between forward and
reverse order)?}
}
\value{
The length of the longest sequence the user correctly repeated.
}
\description{
Does a memory span test using the R console.
}
